// frontend/src/pages/WMS/reports/stockriteria/types/detailStockWms.types.ts
export type TDetailFilter = {
  label?: string;
  prin_code: string;
  productGroupFrom: string;
  productGroupTo: string;
  brand_code_from: string;
  brand_code_to: string;
  prod_code_from: string;
  prod_code_to: string;
  mfg_date_to: Date;
  expirationDateTo: Date;
  mfg_date_from: Date;
  expirationDateFrom: Date;
  txn_date_from?: Date;
  txn_date_to?: Date;
};

export type TDetailReport = {
  company_code?: string;
  prin_code?: string;
  prin_name?: string;
  products?: TDetailProducts[];
};
export type TDetailProducts = {
  prod_code?: string;
  prod_name?: string;
  prod_uppp?: number;
  uom_count?: number;
  bar_code?: string;
  job_no?: string;
  txn_type?: string;
  txn_date?: string;
  key_number?: string;
  packdet_no?: number;
  site_code?: string;
  location_code?: string;
  qty_rcvd?: string;
  qty_stock?: string;
  qty_picked?: string;
  qty_avl?: string;
  pqty_stock?: string;
  lqty_stock?: string;
  pqty_picked?: string;
  lqty_picked?: string;
  pqty_avl?: string;
  lqty_avl?: string;
  p_uom?: string;
  l_uom?: string;
  uppp?: number;
  upp?: number;
  pack_key?: any;
  vessel_name?: string;
  container_no?: string;
  seal_no?: any;
  po_no?: any;
  bl_no?: any;
  doc_ref?: string;
  lot_no?: any;
  pallet_id?: any;
  pallet_serial_no?: any;
  manu_code?: string;
  mfg_date?: any;
  exp_date?: any;
  curr_code?: string;
  ex_rate?: string;
  unit_price?: string;
  freeze_flag?: string;
  freeze_userid?: any;
  freeze_date?: any;
  freeze_reason?: any;
  identity_number?: number;
  user_id?: string;
  user_dt?: string;
  volume?: string;
  net_wt?: string;
  group_code?: string;
  brand_code?: string;
  receipt_type?: any;
  site_ind?: string;
  container_size?: any;
  hs_code?: string;
  qty_moved?: string;
  origin_country?: string;
  barcode?: string;
  model_number?: string;
  cust_code?: any;
  shelf_life_days?: any;
  shelf_life_date?: any;
  prod_attrib_code?: any;
  prod_grade1?: any;
  prod_grade2?: any;
  avg_consumption?: string;
  net_volume?: string;
  batch_no?: any;
  dept_code?: any;
  gross_wt?: string;
  hazmat_ind?: string;
  hazmat_class?: string;
  food_ind?: string;
  pharma_ind?: string;
  special_instructions?: string;
  ref_prod_code?: string;
  batch_type?: number;
  class?: string;
  wob?: string;
  unified_code?: string;
  isTotal?: boolean;
  stockTotal?: boolean;
};

export type TSummary = {
  prin_code: string;
  prod_code: string;
  prod_name: string;
  open_stk_puom: string;
  open_stk_luom: string;
  open_stk: string;
  qty_in_puom: string;
  qty_in_luom: string;
  qty_in: string;
  qty_out_puom: string;
  qty_out_luom: string;
  qty_out: string;
  balance_puom: number;
  balance_luom: number;
  balance: number;
};

export type TSummarydetail = {
  prin_name?: string;
  products?: TSummary[];
};
export type TAgingFilter = {
  label?: string;
  prin_code: string;
  prod_code_from: string;
  prod_code_to: string;
  txn_date_from?: Date;
  txn_date_to?: Date;
  age_1: number;
  age_2: number;
  age_3: number;
  age_4: number;
  age_5: number;
};
export type TAgingReport = {
  COMPANY_CODE: string;
  PRIN_CODE: string;
  PRIN_NAME: string;
  d_age1: number;
  d_age2: number;
  d_age3: number;
  d_age4: number;
  d_age5: number;
  data: TAgingData[];
};
type TAgingData = {
  BELOW_AGE1: string;
  AGE_1: string;
  AGE_2: string;
  AGE_3: string;
  AGE_4: string;
  ABOVE_AGE5: string;
  total: number;
  isAgeing: boolean;
  grand_total: number;
  BELOW_AGE1_Total: number;
  AGE_1_total: number;
  AGE_2_total: number;
  AGE_3_total: number;
  AGE_4_total: number;
  ABOVE_AGE5_total: number;
};